refactor(mobile): split avatar error text
