+++ mobile/lib/features/profile/animated_avatar_capture/error_text.dart
@@ -1,0 +1,22 @@
+part of '../animated_avatar_capture.dart';
+
+class _ErrorText extends StatelessWidget {
+  const _ErrorText(this.message);
+
+  final String message;
+
+  @override
+  Widget build(BuildContext context) => Padding(
+    padding: const EdgeInsets.only(top: Grid.xs),
+    child: Semantics(
+      liveRegion: true,
+      child: Text(
+        message,
+        textAlign: TextAlign.center,
+        style: context.textTheme.bodySmall?.copyWith(
+          color: context.colors.error,
+        ),
+      ),
+    ),
+  );
+}
